+++ src/layouts/AccessWalletLayout/components/NetworkAndAddressModal/NetworkAndAddressModal.vue
@@ -92,17 +92,22 @@
               <div class="dropdown-button-container">
                 <b-dropdown
                   id="hd-derivation-path"
-                  :text="getPathLabel(selectedPath)"
+                  :text="`${getPathLabel(selectedPath)} (${selectedPath}`"
                   right
                   class="dropdown-button-2"
                 >
                   <b-dropdown-item
                     v-for="(val, key) in availablePaths"
-                    :class="selectedPath === val.path ? 'active' : ''"
+                    :class="[
+                      selectedPath === val.path ? 'active' : '',
+                      'dropdown-paths'
+                    ]"
                     :key="'base' + key"
                     @click="changePath(key)"
-                    >{{ val.label }}</b-dropdown-item
                   >
+                    {{ val.label }}
+                    <span>({{ val.path }})</span>
+                  </b-dropdown-item>
                   <b-dropdown-divider />
                   <b-dropdown-item>{{
                     $t('accessWallet.path.custom')
@@ -113,7 +118,9 @@
                     :key="key"
                   >
                     <div class="custom-networks">
-                      <div @click="changePath(key)">{{ val.label }}</div>
+                      <div @click="changePath(key)">
+                        {{ val.label }} <span>({{ val.path }})</span>
+                      </div>
                       <span>
                         <i
                           class="fa fa-times-circle"
@@ -498,5 +505,15 @@ export default {
 
 .activeConn {
   color: gray;
+}
+</style>
+
+<style lang="scss">
+.dropdown-paths {
+  .dropdown-item {
+    align-items: center;
+    display: flex !important;
+    justify-content: space-between;
+  }
 }
 </style>